+++ src/main/java/com/sora/ryokka/repository/ImageRepository.java
@@ -7,8 +7,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ImageRepository extends JpaRepository<ProjectImage, Integer> {
-    List<ProjectImage> findByProjectProjectId(Integer project_projectId);
+public interface ImageRepository extends JpaRepository<ProjectImage, Long> {
+    List<ProjectImage> findByProjectProjectId(Long project_projectId);
 
     ProjectImage findByProjectAndIsMainImageTrue(Project project);
 }
