+++ src/main/java/com/sora/ryokka/repository/ProjectRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProjectRepository extends JpaRepository<Project, Integer> {
-    // Custom query methods can be defined here if needed
+public interface ProjectRepository extends JpaRepository<Project, Long> {
 }
